+++ monkey/infection_monkey/custom_types.py
