+++ monkey/infection_monkey/master/exploiter.py
@@ -7,9 +7,9 @@ from queue import Queue
 from threading import Event
 from typing import Callable, Dict, List, Mapping
 
+from infection_monkey.custom_types import PropagationCredentials
 from infection_monkey.i_puppet import ExploiterResultData, IPuppet
 from infection_monkey.model import VictimHost
-from infection_monkey.typing import PropagationCredentials
 from infection_monkey.utils.threading import interruptible_iter, run_worker_threads
 
 QUEUE_TIMEOUT = 2
